Refatora a função anônima utilizada no forEach da função "getAccountsWithDuplicatePasswords": - Extrai a função "getDecryptedAccountPassword",  que decifra a senha criptografada de uma conta salva. Como se trata de uma ação muito comum, a refatoração também foi aplicada nas demais ocorrências da função. - Altera a lógica de registrar as contas com senhas duplicadas para utilizar somente um Objeto da forma senha -> [contas que têm a senha], deixando a função ciclomaticamente mais simples.
